# Optimize the use of astropy.units