@abhisrkckl 2023-09-23
Current performance comparison before and after these changes (updated periodically):

(The benchmarks use example data for NGC6440E.)

**Before:** 
Chisq computation: 14 ms
Design matrix computation: 25 ms

**After:** 
Chisq computation: 13 ms
Design matrix computation: 11 ms