@dlakaplan 2023-09-28
Is your override function fully accurate?  It's basically the linear approximation, but there are cases where higher-order terms are needed (like for very large proper motions and long time spans).   Do you have any tests to check the accuracy?