@dlakaplan 2023-09-28
As a quick illustration, here is the linear pm approximation vs. what astropy does:
<img width="580" alt="Screen Shot 2023-09-28 at 12 59 53 PM" src="https://github.com/nanograv/PINT/assets/5092062/711bff8a-896f-423d-b0b0-6bf0f28d7cf1">
```
import numpy as np
from astropy.coordinates import SkyCoord
from astropy import units as u, constants as c
from astropy.time import Time
from matplotlib import pyplot as plt

# astropy example
cc = SkyCoord(
    l=10 * u.degree,
    b=45 * u.degree,
    distance=100 * u.pc,
    pm_l_cosb=34 * u.mas / u.yr,
    pm_b=-117 * u.mas / u.yr,
    frame="galactic",
    obstime=Time("1988-12-18 05:11:23.5"),
)

t = cc.obstime + np.linspace(0, 20) * u.yr
ccnew = cc.apply_space_motion(new_obstime=t)

# linear version
l0, b0 = cc.l, cc.b
pml, pmb = cc.pm_l_cosb, cc.pm_b
t0 = cc.obstime
dt = t - t0
l = l0 + pml * dt / np.cos(b0)
b = b0 + pmb * dt

plt.clf()
plt.plot((t - t0).jd, ccnew.l.value - l.value, label="astropy $l$ - linear")
plt.plot((t - t0).jd, ccnew.b.value - b.value, label="astropy $b$ - linear")
plt.xlabel("$t-t_0$ (d)")
plt.ylabel("Position difference (deg)")
plt.legend()
```
the differences are small, but get worse near the pole and scale with the proper motions (and are quadratic in time as expected).