@dlakaplan 2023-09-28
If you look at:
https://docs.astropy.org/en/stable/_modules/astropy/coordinates/sky_coordinate.html#SkyCoord.apply_space_motion
you can see how astropy does it.  it might be possible to use the same method under the hood:
https://github.com/liberfa/erfa/blob/master/src/starpm.c
which is compiled C code so should be fast.  but if you remove/change all of the astropy status checks (which are probably redundant)