@dlakaplan 2023-09-28
So I'm testing just the ERFA function alone, and it seems slower.  so maybe it's not that simple.

It's also possible that the differences between the linear and correct versions depend on things like an accurate distance and RV, which in general we don't have.  So maybe we don't worry.