@dlakaplan 2023-09-28
Another possibility is that if you only need the cartesian components you could do something like:
```

cccartesian = cc.cartesian
x0, y0, z0 = cccartesian.x, cccartesian.y, cccartesian.z
xdot0, ydot0, zdot0 = (
    cccartesian.differentials["s"].d_x.to(1 / u.yr),
    cccartesian.differentials["s"].d_y.to(1 / u.yr),
    cccartesian.differentials["s"].d_z.to(1 / u.yr),
)
xnew = x0 + xdot0 * (t - t0)
ynew = y0 + ydot0 * (t - t0)
znew = z0 + zdot0 * (t - t0)
```