@dlakaplan 2023-09-28
OK, it's not the ERFA function itself that is slow.  It's wrapping it into another `SkyCoord`.  So if you just need the bare RA,Dec and or x,y,z you can do:
```
def erfa_pm_to_cartesian(cc, t):
    icrsrep = cc.icrs.represent_as(
        astropy.coordinates.SphericalRepresentation,
        astropy.coordinates.SphericalDifferential,
    )
    icrsvel = icrsrep.differentials["s"]

    starpm = erfa.pmsafe(
        icrsrep.lon.radian,
        icrsrep.lat.radian,
        icrsvel.d_lon.to_value(u.radian / u.yr),
        icrsvel.d_lat.to_value(u.radian / u.yr),
        0.0,
        0.0,
        cc.obstime.jd1,
        cc.obstime.jd2,
        t.jd1,
        t.jd2,
    )
    ra, dec = u.Quantity(starpm[0], u.radian, copy=False), u.Quantity(
        starpm[1], u.radian, copy=False
    )
    x = np.cos(ra) * np.cos(dec)
    y = np.sin(ra) * np.cos(dec)
    z = np.sin(dec)
    return x, y, z
```
That is faster even than the linear version (maybe because of unit conversion?) and a lot faster than the astropy version.  But I think it agrees exactly with the astropy version.