@dlakaplan 2023-09-28
Comparing:
```
def linear_pm_to_cartesian(cc, t):
    # linear version
    ra0, dec0 = cc.ra, cc.dec
    pmra, pmdec = cc.pm_ra_cosdec, cc.pm_dec
    t0 = cc.obstime
    dt = t - t0
    ra = ra0 + pmra * dt / np.cos(dec0)
    dec = dec0 + pmdec * dt
    x = np.cos(ra) * np.cos(dec)
    y = np.sin(ra) * np.cos(dec)
    z = np.sin(dec)
    return x, y, z


def erfa_pm_to_cartesian(cc, t):
    icrsrep = cc.icrs.represent_as(
        astropy.coordinates.SphericalRepresentation,
        astropy.coordinates.SphericalDifferential,
    )
    icrsvel = icrsrep.differentials["s"]

    starpm = erfa.pmsafe(
        icrsrep.lon.radian,
        icrsrep.lat.radian,
        icrsvel.d_lon.to_value(u.radian / u.yr),
        icrsvel.d_lat.to_value(u.radian / u.yr),
        0.0,
        0.0,
        cc.obstime.jd1,
        cc.obstime.jd2,
        t.jd1,
        t.jd2,
    )
    ra, dec = u.Quantity(starpm[0], u.radian, copy=False), u.Quantity(
        starpm[1], u.radian, copy=False
    )
    x = np.cos(ra) * np.cos(dec)
    y = np.sin(ra) * np.cos(dec)
    z = np.sin(dec)
    return x, y, z


def full_pm_to_cartesian(cc, t):
    ccnew = cc.apply_space_motion(new_obstime=t)
    x = np.cos(ccnew.ra) * np.cos(ccnew.dec)
    y = np.sin(ccnew.ra) * np.cos(ccnew.dec)
    z = np.sin(ccnew.dec)
    return x, y, z
```
I find:
```
Linear: 0.056377207999958046
ERFA: 0.021196374999817635
Full/Astropy: 0.28906691700012743
```