@abhisrkckl 2023-10-04
Closing this because this only implements the linear approximation to the proper motion. A more rigorous treatment is needed for this.